create abstract base wallet, let extension wallet extends it
